+++ paddlenlp/transformers/deepseek_v2/modeling_auto.py
@@ -17,7 +17,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Paddle DeepSeek_V2 model."""
+"""Paddle DeepSeek model."""
 
 from __future__ import annotations
 
@@ -117,13 +117,13 @@ def scaled_dot_product_attention(
         )
 
         if isinstance(outputs, tuple):
-            outputs[0] = outputs[0].reshape([bsz, q_len, v_num_heads, head_dim])
+            outputs[0] = outputs[0].reshape([bsz, kv_seq_len, v_num_heads, head_dim])
             outputs[0] = outputs[0][..., :v_head_dim]
-            outputs[0] = outputs[0].reshape([bsz, q_len, -1])
+            outputs[0] = outputs[0].reshape([bsz, kv_seq_len, -1])
         else:
-            outputs = outputs.reshape([bsz, q_len, v_num_heads, head_dim])
+            outputs = outputs.reshape([bsz, kv_seq_len, v_num_heads, head_dim])
             outputs = outputs[..., :v_head_dim]
-            outputs = outputs.reshape([bsz, q_len, -1])
+            outputs = outputs.reshape([bsz, kv_seq_len, -1])
         return outputs
 
     else:
@@ -170,7 +170,7 @@ def scaled_dot_product_attention(
 
 
 class DeepseekV2MLPAuto(nn.Layer):
-    def __init__(self, config: DeepseekV2Config, hidden_size=None, intermediate_size=None):
+    def __init__(self, config: DeepseekV2Config, hidden_size=None, intermediate_size=None, is_moe=False):
         super().__init__()
         self.config = config
         self.hidden_size = config.hidden_size if hidden_size is None else hidden_size
@@ -217,7 +217,7 @@ class DeepseekV2MoEAuto(MoELayer):
         self.alpha = config.aux_loss_alpha
         if config.n_shared_experts is not None:
             intermediate_size = config.moe_intermediate_size * config.n_shared_experts
-            self.shared_experts = DeepseekV2MLPAuto(config=config, intermediate_size=intermediate_size)
+            self.shared_experts = DeepseekV2MLPAuto(config=config, intermediate_size=intermediate_size, is_moe=True)
 
     def forward(self, hidden_states):
         final_hidden_states, l_aux, l_zloss = super().forward(hidden_states)
@@ -389,13 +389,13 @@ class DeepseekV2AttentionAuto(nn.Layer):
         q_pe, k_pe = apply_rotary_pos_emb(q_pe, k_pe, cos, sin, position_ids)
 
         query_states = paddle.empty([bsz, q_len, self.num_heads, self.q_head_dim], dtype=self.config.dtype)
-        query_states = paddle.concat([q_nope, q_pe], axis=-1)
+        query_states = paddle.concat([q_nope, q_pe], axis=3)
         # query_states[:, :, :, : self.qk_nope_head_dim] = q_nope
         # query_states[:, :, :, self.qk_nope_head_dim :] = q_pe
 
         key_states = paddle.empty([bsz, q_len, self.num_heads, self.q_head_dim], dtype=self.config.dtype)
         # input[0]'s shape = [1, 2048, 16, 128], input[1]'s shape = [1, 2048, 1, 64].
-        key_states = paddle.concat([k_nope, k_pe.expand([bsz, q_len, self.num_heads, k_pe.shape[-1]])], axis=-1)
+        key_states = paddle.concat([k_nope, k_pe.expand([bsz, q_len, self.num_heads, k_pe.shape[-1]])], axis=3)
 
         # key_states[:, :, :, : self.qk_nope_head_dim] = k_nope
         # key_states[:, :, :, self.qk_nope_head_dim :] = k_pe
@@ -973,20 +973,15 @@ class DeepseekV2ForCausalLMAuto(DeepseekV2PretrainedModelAuto):
         if prefix != "":
             assert prefix.endswith(".")
         config = {
-            "dp_config": {"sharding_level": 1, "offload": False, "exclude_layer": None},
             "mp_config": {
                 "parallelize_plan": {
                     f"{prefix}deepseek_v2.embed_tokens": dist.ColWiseParallel(gather_output=True),
-                    f"{prefix}deepseek_v2.layers.*.self_attn.q_b_proj": dist.ColWiseParallel(),
                     f"{prefix}deepseek_v2.layers.*.self_attn.q_proj": dist.ColWiseParallel(),
                     f"{prefix}deepseek_v2.layers.*.self_attn.kv_b_proj": dist.ColWiseParallel(),
                     f"{prefix}deepseek_v2.layers.*.self_attn.o_proj": dist.RowWiseParallel(),
                     f"{prefix}deepseek_v2.layers.*.mlp.gate_proj": dist.ColWiseParallel(),
                     f"{prefix}deepseek_v2.layers.*.mlp.up_proj": dist.ColWiseParallel(),
                     f"{prefix}deepseek_v2.layers.*.mlp.down_proj": dist.RowWiseParallel(),
-                    f"{prefix}deepseek_v2.layers.*.mlp.shared_experts.gate_proj": dist.ColWiseParallel(),
-                    f"{prefix}deepseek_v2.layers.*.mlp.shared_experts.up_proj": dist.ColWiseParallel(),
-                    f"{prefix}deepseek_v2.layers.*.mlp.shared_experts.down_proj": dist.RowWiseParallel(),
                     f"{prefix}lm_head.weight": dist.ColWiseParallel(),
                 }
             },
